+++ CHANGELOG.rst
@@ -3,7 +3,8 @@ Version 0.4 - Development
 * Overall refactoring of API and therefore incompatible with previous versions.
 * Development status set to ``Beta``.
 * Support pypy_ and python 3.3.
-* Added the new :mod:`pulsar.utils.httpurl` module for HTTP tools and HTTP clients.
+* Added the new :mod:`pulsar.utils.httpurl` module for HTTP tools and HTTP 
+  synchronous and asynchronous clients.
 * Refactored :class:`pulsar.Deferred` to be more compatible with twisted. You
   can add separate callbacks for handling errors.
 * Added :class:`pulsar.MultiDeferred` for handling a group of asynchronous
@@ -14,7 +15,7 @@ Version 0.4 - Development
   a dictionary.
 * Socket and WSGI :class:`pulsar.Application` are built on top of the new
   :class:`pulsar.AsyncSocketServer` framework class.
-* **271 regression tests**, **79% coverage**.
+* **279 regression tests**, **80% coverage**.
 
 Version 0.3 - 2012-May-03
 ============================